refactor(app): move InLocalStore to shared/vectors/store refactor(documents): move PineconeStore to shared/vectors feat(vectors): add InLocalStore for local storage feat(vectors): add IndexedDBStore for browser storage feat(vectors): add PineconeStore for pinecone vector db
